number field switched for parseFloat

+++ packages/fluent-controls/src/components/Field/NumberField.tsx
@@ -81,7 +81,7 @@ export const NumberField: React.StatelessComponent<NumberFieldProps> = (props: N
                 postfixClassName={props.postfixClassName}
                 error={!!props.error}
                 disabled={props.disabled}
-                onChange={value => props.onChange(value ? parseInt(value) : null) }
+                onChange={value => props.onChange(value ? parseFloat(value) : null) }
                 className={props.inputClassName}
                 autoFocus={props.autoFocus}
                 attr={{
